requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
dnspython==2.3.0
email-validator==1.3.0
et_xmlfile==2.0.0
faiss-cpu==1.7.3
filelock==3.18.0
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
fsspec==2025.3.2
greenlet==1.1.3.post0
gunicorn==20.1.0
huggingface-hub==0.30.2
idna==3.4
importlib_metadata==8.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
lxml==5.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
pillow==11.2.1
psycopg2-binary==2.9.6
Pygments==2.19.1
pyparsing==3.2.3
pypdf==3.12.0
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==12.6.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
SQLAlchemy==1.4.46
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.5.0
urllib3==2.4.0
Werkzeug==2.2.3
zipp==3.21.0
